mvp/item/pyproject.toml
[project]
name = "item-mvp"
version = "0.1.0"
description = "Item-only MVP for demand forecasting platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.0",
]
